move classes to a package
