+++ tr069/src/test/java/com/github/freeacs/tr069/methods/decision/GetScriptVersionTest.java
@@ -2,6 +2,7 @@ package com.github.freeacs.tr069.methods.decision;
 
 import com.github.freeacs.base.ACSParameters;
 import com.github.freeacs.tr069.CPEParameters;
+import com.github.freeacs.tr069.methods.decision.GetParameterValues.GetScriptVersion;
 import com.github.freeacs.tr069.xml.ParameterValueStruct;
 import org.junit.Test;
